Download Kubernetes binaries for workers, but only if they are not there yet.

+++ kube-bins.sh
@@ -4,9 +4,14 @@ for f in \
   kube-apiserver \
   kube-controller-manager \
   kube-scheduler \
-  kubectl
+  kubectl \
+  kube-proxy \
+  kubelet
 do
   bin=tmp/${f}
-  wget --show-progress --https-only -O ${bin} "https://storage.googleapis.com/kubernetes-release/release/v1.10.1/bin/linux/amd64/${f}"
-  chmod +x ${bin}
+  if [ \! -f ${bin} ]
+  then
+    wget --show-progress --https-only -O ${bin} "https://storage.googleapis.com/kubernetes-release/release/v1.10.1/bin/linux/amd64/${f}"
+    chmod +x ${bin}
+  fi
 done
